War Card Game
A simple Card game where we have assets of all the cards numbered 2 to 14, representing 2 to ace. The design portion was done using Zstacks, Vstacks, and Hstacks, images, text, and spaces. We Then choose a random number from 2-14 and change the state variables based on it. This means every time the user clicks on the button, random numbers are generated, and we use them to replace the card assets. Then we simply increment the scores using if statements based on the random numbers chosen.