The Akpsi Alumni App
An app that pull data from a realtime database that is linked to a google sheets. It then formats that data in a list view which navigates to an expanded view if clicked. The list can be filtered for specific information. It also makes use of the google sign in API to authenticate users.